% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\docType{data}
\name{cells}
\alias{cells}
\title{Single-cell size and geometry data}
\format{
A data frame with 1242 rows and 7 variables:
\describe{
\item{culture_id}{Conditions in which the cell was grown (details provided in \code{\link{cultures}}).}
\item{width}{Width in μm, averaged along the tip-to-tip axis.}
\item{length}{Cell length in μm.}
\item{surface}{Cell surface area in μm^2.}
\item{volume}{Cell volume in μm^3.}
}
}
\usage{
cells
}
\description{
Cell size and geometry data (width, length, surface area and volume) for
1242 \emph{Schizosaccharomyces pombe} cells in base (EMM2) growth medium and 24
modulated growth conditions: decreased glucose concentration, decreased
ammonium concentration, different nitrogen source, or addition of sublethal
doses of the translational inhibitor cycloheximide.
}
\keyword{datasets}
