% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\docType{data}
\name{summaries}
\alias{summaries}
\title{Summary size and geometry data}
\format{
A data frame with 25 rows and loads of variables:
\describe{
\item{culture_id}{Unique culture identifier.}
\item{modulation}{Class of modulation applied to the base medium.}
\item{mod_value}{Quantification of modulation.}
\item{growth_rate}{Population growth rate (h^-1) as measured in turbidostat.}
\item{n_cells}{Number of cells used to calculate summary statistics.}
\item{width_mean, width_sd, length_mean, length_sd, surface_mean, surface_sd, volume_mean, volume_sd, SV_ratio_mean, SV_ratio_sd}{Average and standard deviation of five size measures: average cell width along tip-to-tip axis (μm), cell length (μm), cell surface area (μm^2), cell volume (μm^3), and surface-to-volume ratio (μm^-1).}
}
}
\usage{
summaries
}
\description{
Means and standard deviations of single-cell size data from
\code{\link{cells}}. Some information about the growth media is provided, see
\code{\link{cultures}} for more details.
}
\keyword{datasets}
